pyproject.toml
[tool.poetry]
name = "repl_python3_beta_AuskaBot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
bs4 = "^0.0.1"
youtube_dl = "^2021.6.6"
PyNaCl = "^1.4.0"
Flask = "^2.0.1"
animec = "^0.3.3"
python-decouple = "^3.4"
colorama = "^0.4.4"
config = "^0.5.1"
"nekos.py" = "^1.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
